requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
faster-whisper
ffmpeg-python
python-multipart